src/requirements.txt
grpcio==1.51.3
protobuf==3.19.4
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.2.1